refactor: extract DeathHandler before removing RestHandler
DeathHandler receives the rollDeathSave + die logic verbatim so Phase D's
rest-feature removal can retire RestHandler cleanly. No behavior change.
RestHandlerTest renamed to DeathHandlerTest; rest-specific test cases removed
since those methods are stubs pending D2 removal.
